refactor(codegen): extrai passes/static_fields.rs de func.rs
Move expand_static_fields + 6 helpers (468 linhas) para
crates/rts-codegen/src/codegen/lower/passes/static_fields.rs:
- default_expr_for_ann, build_type_ann
- rewrite_static_in_program / _statement / _swc_stmt / _expr

Reescreve `static x = init` de classe pra `let __class_static_C_x`
top-level e propaga `C.x` -> Ident(global) em todo o programa.

func.rs: 7178 -> 6713 linhas. Suite verde.
